requirements.txt
asgiref==3.3.4
boto3==1.17.72
botocore==1.20.72
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
django-ckeditor==6.0.0
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==4.1.1
django-modeltranslation==0.17.1
django-storages==1.11.1
django-taggit==1.4.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
mysqlclient==2.0.3
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
